--- src/main/java/cholog/supp/api/post/dto/PostRequest.java
@@ -1,8 +1,0 @@
-package cholog.supp.api.post.dto;
-
-public record PostRequest(
-    String title,
-    String content,
-    Long studyId) {
-
-}
