+++ src/main/java/cholog/supp/api/post/dto/request/CreatePostRequest.java
@@ -1,0 +1,8 @@
+package cholog.supp.api.post.dto.request;
+
+public record CreatePostRequest(
+    String title,
+    String content,
+    Long studyId) {
+
+}
