+++ src/main/java/cholog/supp/api/study/dto/request/CreateStudyGroupRequest.java
@@ -1,4 +1,4 @@
-package cholog.supp.api.study.dto;
+package cholog.supp.api.study.dto.request;
 
 public record CreateStudyGroupRequest(String studyName, String organization) {
